requirements.txt
aiohttp>=3.9.0
beautifulsoup4==4.12.2
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv